+++ .dfx/local/canister_ids.json
@@ -2,10 +2,7 @@
   "__Candid_UI": {
     "local": "bd3sg-teaaa-aaaaa-qaaba-cai"
   },
-  "hello_world": {
-    "local": "bkyz2-fmaaa-aaaaa-qaaaq-cai"
-  },
   "skill-swap": {
-    "local": "be2us-64aaa-aaaaa-qaabq-cai"
+    "local": "bkyz2-fmaaa-aaaaa-qaaaq-cai"
   }
 }
