--- .dfx/local/canisters/hello_world/init_args.txt
@@ -1,1 +1,0 @@
-()
