--- .dfx/local/canisters/hello_world/service.did
@@ -1,156 +1,0 @@
-service : {
-  acceptOffer : (record { offerId : principal }) -> (
-      variant {
-        Ok : text;
-        Err : variant {
-          InvalidRating : text;
-          UsernameTaken : text;
-          OfferNotFound : text;
-          OfferAlreadyMatched : text;
-          MatchNotFound : text;
-          InvalidSkillLevel : text;
-          UnauthorizedAction : text;
-          UserNotFound : text;
-        };
-      },
-    );
-  addSkill : (
-      record {
-        experienceLevel : text;
-        name : text;
-        description : text;
-        category : text;
-      },
-    ) -> (
-      variant {
-        Ok : text;
-        Err : variant {
-          InvalidRating : text;
-          UsernameTaken : text;
-          OfferNotFound : text;
-          OfferAlreadyMatched : text;
-          MatchNotFound : text;
-          InvalidSkillLevel : text;
-          UnauthorizedAction : text;
-          UserNotFound : text;
-        };
-      },
-    );
-  completeLesson : (
-      record { feedback : text; matchId : principal; rating : nat64 },
-    ) -> (
-      variant {
-        Ok : text;
-        Err : variant {
-          InvalidRating : text;
-          UsernameTaken : text;
-          OfferNotFound : text;
-          OfferAlreadyMatched : text;
-          MatchNotFound : text;
-          InvalidSkillLevel : text;
-          UnauthorizedAction : text;
-          UserNotFound : text;
-        };
-      },
-    );
-  createOffer : (
-      record {
-        skillWantedDescription : text;
-        skillOfferedLevel : text;
-        skillOfferedCategory : text;
-        skillWantedLevel : text;
-        skillOfferedName : text;
-        skillWantedCategory : text;
-        skillWantedName : text;
-        skillOfferedDescription : text;
-      },
-    ) -> (
-      variant {
-        Ok : text;
-        Err : variant {
-          InvalidRating : text;
-          UsernameTaken : text;
-          OfferNotFound : text;
-          OfferAlreadyMatched : text;
-          MatchNotFound : text;
-          InvalidSkillLevel : text;
-          UnauthorizedAction : text;
-          UserNotFound : text;
-        };
-      },
-    );
-  createUser : (record { username : text }) -> (
-      variant {
-        Ok : text;
-        Err : variant {
-          InvalidRating : text;
-          UsernameTaken : text;
-          OfferNotFound : text;
-          OfferAlreadyMatched : text;
-          MatchNotFound : text;
-          InvalidSkillLevel : text;
-          UnauthorizedAction : text;
-          UserNotFound : text;
-        };
-      },
-    );
-  getActiveOffers : () -> (
-      vec record {
-        id : principal;
-        status : text;
-        createdAt : nat64;
-        teacher : principal;
-        skillWanted : record {
-          experienceLevel : text;
-          name : text;
-          description : text;
-          category : text;
-        };
-        skillOffered : record {
-          experienceLevel : text;
-          name : text;
-          description : text;
-          category : text;
-        };
-      },
-    ) query;
-  getUserMatches : (principal) -> (
-      vec record {
-        id : principal;
-        status : text;
-        offererId : principal;
-        createdAt : nat64;
-        feedback : text;
-        rating : nat64;
-        accepterId : principal;
-        skillOffererId : principal;
-      },
-    ) query;
-  getUserProfile : (principal) -> (
-      variant {
-        Ok : record {
-          id : principal;
-          username : text;
-          dateJoined : nat64;
-          rating : nat64;
-          completedExchanges : nat64;
-          skills : vec record {
-            experienceLevel : text;
-            name : text;
-            description : text;
-            category : text;
-          };
-        };
-        Err : variant {
-          InvalidRating : text;
-          UsernameTaken : text;
-          OfferNotFound : text;
-          OfferAlreadyMatched : text;
-          MatchNotFound : text;
-          InvalidSkillLevel : text;
-          UnauthorizedAction : text;
-          UserNotFound : text;
-        };
-      },
-    ) query;
-}
