+++ .dfx/local/canisters/skill-swap/constructor.did
@@ -21,6 +21,7 @@ service : () -> {
           OfferAlreadyMatched : text;
           MatchNotFound : text;
           InvalidSkillLevel : text;
+          InternalError : text;
           UnauthorizedAction : text;
           UserNotFound : text;
         };
@@ -51,6 +52,7 @@ service : () -> {
           OfferAlreadyMatched : text;
           MatchNotFound : text;
           InvalidSkillLevel : text;
+          InternalError : text;
           UnauthorizedAction : text;
           UserNotFound : text;
         };
@@ -68,6 +70,7 @@ service : () -> {
           OfferAlreadyMatched : text;
           MatchNotFound : text;
           InvalidSkillLevel : text;
+          InternalError : text;
           UnauthorizedAction : text;
           UserNotFound : text;
         };
@@ -114,6 +117,7 @@ service : () -> {
           OfferAlreadyMatched : text;
           MatchNotFound : text;
           InvalidSkillLevel : text;
+          InternalError : text;
           UnauthorizedAction : text;
           UserNotFound : text;
         };
@@ -144,6 +148,7 @@ service : () -> {
           OfferAlreadyMatched : text;
           MatchNotFound : text;
           InvalidSkillLevel : text;
+          InternalError : text;
           UnauthorizedAction : text;
           UserNotFound : text;
         };
@@ -203,6 +208,7 @@ service : () -> {
           OfferAlreadyMatched : text;
           MatchNotFound : text;
           InvalidSkillLevel : text;
+          InternalError : text;
           UnauthorizedAction : text;
           UserNotFound : text;
         };
