+++ .dfx/local/canisters/skill-swap/service.did.js
@@ -24,6 +24,7 @@ export const idlFactory = ({ IDL }) => {
               'OfferAlreadyMatched' : IDL.Text,
               'MatchNotFound' : IDL.Text,
               'InvalidSkillLevel' : IDL.Text,
+              'InternalError' : IDL.Text,
               'UnauthorizedAction' : IDL.Text,
               'UserNotFound' : IDL.Text,
             }),
@@ -58,6 +59,7 @@ export const idlFactory = ({ IDL }) => {
               'OfferAlreadyMatched' : IDL.Text,
               'MatchNotFound' : IDL.Text,
               'InvalidSkillLevel' : IDL.Text,
+              'InternalError' : IDL.Text,
               'UnauthorizedAction' : IDL.Text,
               'UserNotFound' : IDL.Text,
             }),
@@ -83,6 +85,7 @@ export const idlFactory = ({ IDL }) => {
               'OfferAlreadyMatched' : IDL.Text,
               'MatchNotFound' : IDL.Text,
               'InvalidSkillLevel' : IDL.Text,
+              'InternalError' : IDL.Text,
               'UnauthorizedAction' : IDL.Text,
               'UserNotFound' : IDL.Text,
             }),
@@ -133,6 +136,7 @@ export const idlFactory = ({ IDL }) => {
               'OfferAlreadyMatched' : IDL.Text,
               'MatchNotFound' : IDL.Text,
               'InvalidSkillLevel' : IDL.Text,
+              'InternalError' : IDL.Text,
               'UnauthorizedAction' : IDL.Text,
               'UserNotFound' : IDL.Text,
             }),
@@ -169,6 +173,7 @@ export const idlFactory = ({ IDL }) => {
               'OfferAlreadyMatched' : IDL.Text,
               'MatchNotFound' : IDL.Text,
               'InvalidSkillLevel' : IDL.Text,
+              'InternalError' : IDL.Text,
               'UnauthorizedAction' : IDL.Text,
               'UserNotFound' : IDL.Text,
             }),
@@ -246,6 +251,7 @@ export const idlFactory = ({ IDL }) => {
               'OfferAlreadyMatched' : IDL.Text,
               'MatchNotFound' : IDL.Text,
               'InvalidSkillLevel' : IDL.Text,
+              'InternalError' : IDL.Text,
               'UnauthorizedAction' : IDL.Text,
               'UserNotFound' : IDL.Text,
             }),
