+++ .dfx/local/network-id
@@ -1,4 +1,4 @@
 {
-  "created": "2025-01-01 16:24:26.183145485 +00:00:00",
-  "settings_digest": "658f265fafd9c003683ad2f729e00c1ec43c45139dd9dd08114c24e76a3895f1"
+  "created": "2025-01-17 10:26:38.53976019 +00:00:00",
+  "settings_digest": "a2c7b27b81377d6b5f7e38775792e6f435f034f3afc27553f2f3eed938fb7e9b"
 }
