+++ src/index.ts
@@ -3,138 +3,18 @@ import {
   Err,
   Ok,
   Principal,
-  Record,
   Result,
-  StableBTreeMap,
-  Variant,
   Vec,
   ic,
-  nat64,
   query,
   update,
   text,
 } from "azle/experimental";
-
-// Define core types
-const Skill = Record({
-  name: text,
-  category: text,
-  experienceLevel: text, // "beginner", "intermediate", "expert"
-  description: text,
-});
-
-type Skill = typeof Skill.tsType;
-
-// Skill response type
-const SkillResponse = Record({
-  message: text,
-  skill: Skill,
-});
-
-const SkillOffer = Record({
-  id: Principal,
-  teacher: Principal,
-  skillOffered: Skill,
-  skillWanted: Skill,
-  status: text, // "active", "matched", "completed"
-  createdAt: nat64,
-});
-
-type SkillOffer = typeof SkillOffer.tsType;
-
-// Skill Offer response type
-const CreateOfferResponse = Record({
-  message: text,
-  offer: SkillOffer,
-});
-
-const Match = Record({
-  id: Principal,
-  offererId: Principal,
-  accepterId: Principal,
-  skillOffererId: Principal,
-  status: text, // "ongoing", "completed", "cancelled"
-  rating: nat64,
-  feedback: text,
-  createdAt: nat64,
-});
-
-type Match = typeof Match.tsType;
-
-// Match response type
-const MatchResponse = Record({
-  message: text,
-  match: Match,
-});
-
-const User = Record({
-  id: Principal,
-  username: text,
-  skills: Vec(Skill),
-  rating: nat64,
-  completedExchanges: nat64,
-  dateJoined: nat64,
-});
-
-type User = typeof User.tsType;
-
-// User response type
-const CreateUserResponse = Record({
-  message: text,
-  user: User,
-});
-
-// Define payload types
-const CreateUserPayload = Record({
-  username: text,
-});
-
-const AddSkillPayload = Record({
-  name: text,
-  category: text,
-  experienceLevel: text,
-  description: text,
-});
-
-const CreateOfferPayload = Record({
-  skillOfferedName: text,
-  skillOfferedCategory: text,
-  skillOfferedLevel: text,
-  skillOfferedDescription: text,
-  skillWantedName: text,
-  skillWantedCategory: text,
-  skillWantedLevel: text,
-  skillWantedDescription: text,
-});
-
-const AcceptOfferPayload = Record({
-  offerId: Principal,
-});
-
-const CompleteLessonPayload = Record({
-  matchId: Principal,
-  rating: nat64,
-  feedback: text,
-});
-
-// Error variants
-const Errors = Variant({
-  UserNotFound: text,
-  OfferNotFound: text,
-  MatchNotFound: text,
-  InvalidRating: text,
-  UsernameTaken: text,
-  UnauthorizedAction: text,
-  OfferAlreadyMatched: text,
-  InvalidSkillLevel: text,
-});
-
-type Errors = typeof Errors.tsType;
-
-// Storage
-const userStorage = StableBTreeMap<Principal, typeof User.tsType>(0);
-const offerStorage = StableBTreeMap<Principal, typeof SkillOffer.tsType>(1);
-const matchStorage = StableBTreeMap<Principal, typeof Match.tsType>(2);
+import UserController from "./controllers/UserController";
+import { AcceptOfferPayload, AddSkillPayload, CompleteLessonPayload, CreateOfferPayload, CreateOfferResponse, CreateUserPayload, CreateUserResponse, Errors, Match, MatchResponse, SkillOffer, SkillResponse, User } from "./datatypes/dataType";
+import OfferController from "./controllers/OfferController";
+import { offerStorage } from "./storages/storage";
+import SkillController from "./controllers/SkillController";
 
 export default Canister({
   // User Management
@@ -142,32 +22,7 @@ export default Canister({
     [CreateUserPayload],
     Result(CreateUserResponse, Errors),
     (payload) => {
-      const caller = ic.caller();
-
-      // Check if username exists
-      const existingUser = Array.from(userStorage.values()).find(
-        (user) => user.username === payload.username
-      );
-      if (existingUser) {
-        return Err({
-          UsernameTaken: `Username ${payload.username} is already taken`,
-        });
-      }
-
-      const newUser: User = {
-        id: caller,
-        username: payload.username,
-        skills: [],
-        rating: 0n,
-        completedExchanges: 0n,
-        dateJoined: ic.time(),
-      };
-
-      userStorage.insert(caller, newUser);
-      return Ok({
-        message: `User ${payload.username} created successfully`,
-        user: newUser,
-      });
+      return UserController.createUser(payload)
     }
   ),
 
@@ -176,38 +31,7 @@ export default Canister({
     [AddSkillPayload],
     Result(SkillResponse, Errors),
     (payload) => {
-      const caller = ic.caller();
-      const user = userStorage.get(caller);
-
-      if (!user) {
-        return Err({ UserNotFound: "User not found" });
-      }
-
-      const validLevels = ["beginner", "intermediate", "expert"];
-      if (!validLevels.includes(payload.experienceLevel.toLowerCase())) {
-        return Err({
-          InvalidSkillLevel:
-            "Invalid skill level. Use: beginner, intermediate, or expert",
-        });
-      }
-
-      const newSkill = {
-        name: payload.name,
-        category: payload.category,
-        experienceLevel: payload.experienceLevel.toLowerCase(),
-        description: payload.description,
-      };
-
-      const updatedUser = {
-        ...user,
-        skills: [...user.skills, newSkill],
-      };
-
-      userStorage.insert(caller, updatedUser);
-      return Ok({
-        message: `Skill ${payload.name} added successfully`,
-        skill: newSkill,
-      });
+      return SkillController.addSkill(payload)
     }
   ),
 
@@ -216,41 +40,7 @@ export default Canister({
     [CreateOfferPayload],
     Result(CreateOfferResponse, Errors),
     (payload) => {
-      const caller = ic.caller();
-      const user = userStorage.get(caller);
-
-      if (!user) {
-        return Err({ UserNotFound: "User not found" });
-      }
-
-      const skillOffered = {
-        name: payload.skillOfferedName,
-        category: payload.skillOfferedCategory,
-        experienceLevel: payload.skillOfferedLevel,
-        description: payload.skillOfferedDescription,
-      };
-
-      const skillWanted = {
-        name: payload.skillWantedName,
-        category: payload.skillWantedCategory,
-        experienceLevel: payload.skillWantedLevel,
-        description: payload.skillWantedDescription,
-      };
-
-      const newOffer = {
-        id: generateId(),
-        teacher: caller,
-        skillOffered,
-        skillWanted,
-        status: "active",
-        createdAt: ic.time(),
-      };
-
-      offerStorage.insert(newOffer.id, newOffer);
-      return Ok({
-        message: `Offer created successfully`,
-        offer: newOffer,
-      });
+      return OfferController.createOffer(payload)
     }
   ),
 
@@ -259,41 +49,7 @@ export default Canister({
     [AcceptOfferPayload],
     Result(MatchResponse, Errors),
     (payload) => {
-      const caller = ic.caller();
-      const offer = offerStorage.get(payload.offerId);
-
-      if (!offer) {
-        return Err({ OfferNotFound: "Offer not found" });
-      }
-
-      if (offer.status !== "active") {
-        return Err({ OfferAlreadyMatched: "This offer is no longer active" });
-      }
-
-      const match: Match = {
-        id: generateId(),
-        offererId: offer.teacher,
-        accepterId: caller,
-        skillOffererId: payload.offerId,
-        status: "ongoing",
-        rating: 0n,
-        feedback: "",
-        createdAt: ic.time(),
-      };
-
-      // Update offer status
-      const updatedOffer = {
-        ...offer,
-        status: "matched",
-      };
-
-      matchStorage.insert(match.id, match);
-      offerStorage.insert(payload.offerId, updatedOffer);
-
-      return Ok({
-        message: `Offer accepted successfully`,
-        match,
-      });
+      return OfferController.acceptOffer(payload)
     }
   ),
 
@@ -302,66 +58,23 @@ export default Canister({
     [CompleteLessonPayload],
     Result(text, Errors),
     (payload) => {
-      const caller = ic.caller();
-      const match = matchStorage.get(payload.matchId);
-
-      if (!match) {
-        return Err({ MatchNotFound: "Match not found" });
-      }
-
-      if (payload.rating < 1 || payload.rating > 5) {
-        return Err({ InvalidRating: "Rating must be between 1 and 5" });
-      }
-
-      const updatedMatch = {
-        ...match,
-        status: "completed",
-        rating: payload.rating,
-        feedback: payload.feedback,
-      };
-
-      // Update user stats
-      const teacher = userStorage.get(match.offererId);
-      if (teacher) {
-        const updatedTeacher = {
-          ...teacher,
-          rating: (teacher.rating + payload.rating) / 2n,
-          completedExchanges: teacher.completedExchanges + 1n,
-        };
-        userStorage.insert(match.offererId, updatedTeacher);
-      }
-
-      matchStorage.insert(payload.matchId, updatedMatch);
-      return Ok("Lesson completed and rated successfully");
+      return SkillController.completeLesson(payload)
     }
   ),
 
   // Queries
   getActiveOffers: query([], Vec(SkillOffer), () => {
-    return offerStorage.values().filter((offer) => offer.status === "active");
+    return OfferController.getActiveOffers()
   }),
 
   getUserProfile: query([Principal], Result(User, Errors), (userId) => {
-    const user = userStorage.get(userId);
-    if (!user) {
-      return Err({ UserNotFound: "User not found" });
-    }
-    return Ok(user);
+    return UserController.getUserProfile(userId)
   }),
 
   getUserMatches: query([Principal], Vec(Match), (userId) => {
-    return matchStorage
-      .values()
-      .filter(
-        (match) => match.offererId === userId || match.accepterId === userId
-      );
+    return UserController.getUserMatches(userId)
   }),
 });
 
 // Helper function
-function generateId(): Principal {
-  const randomBytes = new Array(29)
-    .fill(0)
-    .map((_) => Math.floor(Math.random() * 256));
-  return Principal.fromUint8Array(Uint8Array.from(randomBytes));
-}
+
