+++ src/utils/generateId.ts
@@ -1,0 +1,9 @@
+import { Principal } from "azle/experimental";
+
+
+export function generateId(): Principal {
+    const randomBytes = new Array(29)
+        .fill(0)
+        .map((_) => Math.floor(Math.random() * 256));
+    return Principal.fromUint8Array(Uint8Array.from(randomBytes));
+}
